requirements.txt
gdal
rpcm
opencv-contrib-python
jupyter
Pillow
chardet
matplotlib
numpy
affine
fire
kornia
plyflatten
pyproj
pytorch_lightning
torchmetrics
PyYAML
rasterio
scipy
srtm4
utm
scikit_image
numba
